+++ src/app/(component)/Home/home.jsx
@@ -1,4 +1,4 @@
-"use client"; // Enables client-side rendering
+"use client";
 
 import React from "react";
 import Image from "next/image";
@@ -10,20 +10,28 @@ export default function Home() {
 	const { userLoggedIn } = useAuth(); // Get user login status
 	const router = useRouter(); // Access navigation
 
-	// Redirect to homepage if not logged in
-	console.log("is user logged in: " + userLoggedIn);
+	// console.log("is user logged in: " + userLoggedIn);
 
 	return (
 		<div className={styles.container}>
-			{!userLoggedIn && router.push("/")} {/* Redirect if not logged in */}
+			{/* Redirect if not logged in */}
+			{!userLoggedIn && router.push("/")}
 			{/* Text and description section */}
 			<div className={styles.textContainer}>
 				<h1 className={styles.title}>Top platform for stress relief</h1>
-				<p className={styles.desc}>Our app combines journaling and games to help you unwind. Find calm and balance in one place.</p>
+				<p className={styles.desc}>
+					Our app combines journaling and games to help you unwind.
+					Find calm and balance in one place.
+				</p>
 			</div>
 			{/* Image section */}
 			<div className={styles.imgContainer}>
-				<Image src="/logo.svg" alt="Stress Relief Logo" fill className={styles.heroImg} />
+				<Image
+					src="/logo.svg"
+					alt="Stress Relief Logo"
+					fill
+					className={styles.heroImg}
+				/>
 			</div>
 		</div>
 	);
