+++ src/app/(component)/Journal/journal.jsx
