+++ src/app/journal/page.jsx
